+++ src/encoded/__init__.py
@@ -232,6 +232,8 @@ def main(global_config, **local_config):
     config.include('.batch_download')
     config.include('.loadxl')
     config.include('.visualization')
+    config.include('.ingestion_listener')
+    config.include('.ingestion.ingestion_message_handler_default')
 
     if 'elasticsearch.server' in config.registry.settings:
         config.include('snovault.elasticsearch')