+++ CHANGELOG.rst
@@ -6,6 +6,51 @@ fourfront
 Change Log
 ----------
 
+5.2.0
+======
+
+* Changed ``deploy/docker/production/entrypoint.bash`` to include ``entrypoint_ingester``.
+* Added ``deploy/docker/production/entrypoint_ingester.bash``.
+* Added ``src/encoded/ingester/ingestion_listener_base.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/ingestion_message.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/ingestion_message_handler_decorator.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/common.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/exceptions.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/queue_utils.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/ingester/processors.py`` (From cgap-portal except
+  removed ``handle_genelist``, ``handle_variant_update``, ``handle_metadata_bundle``,
+  ``handle_simulated_bundle`` ``simulated_processor`` and added ``handle_ontology_update``
+  which (the latter) is from the ``fourfront`` ``ff_ingester`` branch).
+* Added ``src/encoded/ingestion_listener.py`` (verbatim from cgap-portal).
+* Added ``src/encoded/types/ingestion.py`` (verbatim from cgap-portal).
+* Changed ``src/encoded/utils.py``:
+  * Changed ``print`` to ``PRINT`` throughout. Added ``log``.
+  * Changed ``s3_output_stream`` to add arg (and extra kwargs) for ``s3_encrypt_key_id``.
+  * Added ``extra_kwargs_for_s3_encrypt_key_id`` function (verbatim from cgap-portal).
+  * Added ``SettingsKey`` class (verbatim from cgap-portal).
+  * Added ``ExtraArgs`` class (verbatim from cgap-portal).
+  * Changed ``create_empty_s3_file`` to add arg (and extra kwargs) for ``s3_encrypt_key_id``.
+  * Added ``_app_from_clues`` function (verbatim from cgap-portal).
+  * Added ``make_vapp_for_email`` function (verbatim from cgap-portal).
+  * Added ``vapp_for_email`` function (verbatim from cgap-portal).
+  * Added ``make_vapp_for_ingestion`` function (verbatim from cgap-portal).
+  * Added ``vapp_for_ingestion`` function (verbatim from cgap-portal).
+  * Added ``make_s3_client`` function (verbatim from cgap-portal except log.info not log.warning).
+  * Added ``build_s3_presigned_get_url`` function (verbatim from cgap-portal).
+  * Added ``convert_integer_to_comma_string`` function (verbatim from cgap-portal).
+* Changed ``src/encoded/__init__.py`` to include in ``main``
+  ``config.include('.ingestion_listener')`` and
+  ``config.include('.ingestion.ingestion_message_handler_default')``.
+* Changed ``pyproject.toml`` to
+  add ``ingester = "encoded.ingestion_listener:composite"``
+  to ``[paste.composite_factory]`` section
+  and ``ingestion-listener = "encoded.ingestion_listener:main"``
+  to ``[tool.poetry.scripts]`` section.
+* Changed ``Makefile`` to include in ingestion code (from cgap-portal).
+* Changed ``src/encoded/dev_servers.py`` to include in ingestion code (from cgap-portal).
+* Added ``src/encoded/schemas/ingestion_submission.json`` (verbatim from cgap-portal).
+
+
 5.1.1
 ======